Add checks to BigInteger creation and toInteger
Ensure properly formatted string, when passing a string in. Separate code path for integers and strings, just to be more explicit in types.

Throws an exception if receiving an invalid string to the create method.

Throws an exception if converting to integer, and the value can not fit in a PHP integer.